Doorkeeper

+++ app/controllers/api/v1/mechanics_controller.rb
@@ -1,5 +1,6 @@
 class Api::V1::MechanicsController < Api::V1::BaseController
-      def index
+    before_action :doorkeeper_authorize!
+    def index
         if params[:q].present?
             mechanics = Mechanic.where("name ILIKE ?", "%#{params[:q]}%")
         else
@@ -8,62 +9,23 @@ class Api::V1::MechanicsController < Api::V1::BaseController
         render json: mechanics
     end
 
-    def new
-        mechanic = Mechanic.new
-    end
-
-    def create
-        mechanic = Mechanic.new(mechanic_params)
-
-        if mechanic.save
-            if params[:return_to].present?
-                redirect_to params[:return_to]
-            else
-                redirect_back(fallback_location: :records_path)
-            end
-            render json: { message: "Created successfully",
-              mechanic: mechanic
-            }, status: :ok
-        else
-            render json:{
-              error: mechanic.errors.full_messages
-            }, status: :unprocessable_entity
-        end
-    end
-
-    def show
-        mechanic = Mechanic.find(params[:id])
-        render json: mechanic
-    end
-
     def update
-        mechanic = Mechanic.find(params[:id])
-
-        if mechanic.update(params[:mechanic].permit(:name, :email, :experience))
-            redirect_to profile_show_path(mechanic), notice: "Mechanic updated successfully"
-        else
-            flash.now[:alert] = mechanic.errors.full_messages.join(", ")
-            render :edit, status: :unprocessable_entity
-        end
-    end
-
-    def mechanic_params
-        params.require(:mechanic).permit(:name, :experience)
-    end
-
-    def destroy
-        mechanic = Mechanic.find(params[:id])
-        mechanic.destroy
-        redirect_to mechanics_path, notice: 'Mechanic was successfully deleted.'
-    end
-
-    def disable
         mechanic = Mechanic.find_by(id: params[:id])
         if mechanic
-        mechanic.lock_access!
-        redirect_to root_path, notice: "Mechanic disabled successfully."
+            if mechanic&.update(params[:mechanic].permit(:name, :email, :experience))
+                render json: {
+                    message: "Updated successfully",
+                    mechanic: mechanic
+                }, status: :ok
+            else
+                render json:{
+                    error: mechanic.errors.full_messages.join(", ")
+                }, status: :unprocessable_entity
+            end
         else
-        redirect_to profile_show_path(mechanic), alert: "Mechanic not found."
+            render json: {
+                    error: "Mechanic not found"
+                }, status: :not_found
         end
     end
 
@@ -88,23 +50,8 @@ class Api::V1::MechanicsController < Api::V1::BaseController
     end
 
     def show
-        mechanic = Mechanic.find(params[:id])
+        mechanic = Mechanic.find_by(id: params[:id])
         render json: mechanic
-    end
-
-    def update
-        mechanic = Mechanic.find(params[:id])
-
-        if mechanic.update(params[:mechanic].permit(:name, :experience))
-          render json: {
-            message: "Updated successfully",
-            mechanic: mechanic
-          }, status: :ok
-        else
-            render json: {
-              error: mechanic.errors.full_messages.join(", ")
-            }, status: :unprocessable_entity
-        end
     end
 
     def mechanic_params
@@ -112,24 +59,43 @@ class Api::V1::MechanicsController < Api::V1::BaseController
     end
 
     def destroy
-      mechanic = Mechanic.find(params[:id])
-
-      if mechanic.destroy
-        render json: { message: "Mechanic deleted successfully" }, status: :ok
-      else
-        render json: { errors: mechanic.errors.full_messages }, status: :unprocessable_entity
-      end
+      mechanic = Mechanic.find_by(id: params[:id])
+        if mechanic
+            if mechanic.destroy
+                render json: { message: "Mechanic deleted successfully" }, status: :ok
+            else
+                render json: { errors: mechanic.errors.full_messages }, status: :unprocessable_entity
+            end
+        else
+            render json: { error: "Mechanic not found."
+        }, status: :not_found
+        end
     end
 
     def disable
       mechanic = Mechanic.find_by(id: params[:id])
       if mechanic
         mechanic.lock_access!
-        render json: { message: "Disabled successfully",
+        render json: { message: "Disabled successfully"
       }, status: :ok
       else
-        render json: { error: "Mechanic not found.",
+        render json: { error: "Mechanic not found."
         }, status: :not_found
       end
+    end
+
+    def unlock
+    mechanic = Mechanic.find_by(id: params[:id])
+    puts mechanic
+        if mechanic
+            if mechanic.access_locked?
+                mechanic.unlock_access!
+            end
+            render json: { message: "Unlocked successfully"
+            }, status: :ok
+        else
+            render json: { error: "Mechanic not found."
+            }, status: :not_found
+        end
     end
 end
